python_service/requirements_api.txt
# Flask API 服务依赖
Flask==3.0.0
flask-cors==4.0.0
tensorflow>=2.20.0,<3.0.0
numpy>=2.3.2,<3.0.0
Pillow>=11.3.0,<12.0.0
openai>=1.0.0

